Finish moving from old net package to new net package for #23
